pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memory-vault"
version = "0.1.0-alpha"
description = "Secure, owner-sovereign, offline-first storage for high-value cognitive artifacts"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "kase1111-hash"}
]
keywords = [
    "security",
    "encryption",
    "memory",
    "vault",
    "ai-agents",
    "privacy",
    "cryptography",
    "offline-first"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Archiving",
]
requires-python = ">=3.7"
dependencies = [
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
tpm = [
    "tpm2-pytss>=2.1.0; platform_system=='Linux'",
]
tokens = [
    "fido2>=1.1.0",
    "pyotp>=2.8.0",
]
all = [
    "tpm2-pytss>=2.1.0; platform_system=='Linux'",
    "fido2>=1.1.0",
    "pyotp>=2.8.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "pip-audit>=2.6.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
memory-vault = "cli:main"

[project.urls]
Homepage = "https://github.com/kase1111-hash/memory-vault"
Documentation = "https://github.com/kase1111-hash/memory-vault#readme"
Repository = "https://github.com/kase1111-hash/memory-vault.git"
Issues = "https://github.com/kase1111-hash/memory-vault/issues"

[tool.setuptools]
py-modules = ["vault", "crypto", "db", "models", "merkle", "cli", "boundry", "agent_os", "physical_token", "deadman", "intentlog", "natlangchain", "effort", "escrow", "zkproofs", "errors", "siem_reporter"]
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "*/__pycache__/*", "setup.py", "conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py38"
line-length = 120
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "B",      # flake8-bugbear
    "S",      # flake8-bandit (security)
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # assert usage (needed for tests)
    "S311",   # pseudo-random (acceptable for non-crypto uses)
    "B008",   # function call in default argument
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Re-exports for public API
"tests/*" = ["S101", "S105", "S106", "S108"]  # Allow asserts, hardcoded passwords, and temp paths in tests
